requirements.txt
# general modules
django==5.2.7
python-decouple==3.8
pillow==12.0.0
djangorestframework


# third party modules
markdown      
django-filter  
urllib3==2.5.0
coreapi==2.3.3
coreschema==0.0.4
drf-yasg[validation]
djangorestframework-simplejwt
django-cors-headers


# deployment modules

# email third party modules
django-mail-templated

# refactoring and tests
black 
flake8
pytest
pytest-django
Faker

# Background process and cache
celery
redis
